refactor : 멤버 관리 컨트롤러(+테스트) 분리
